implemented all unit test

+++ src/service/users/users.service.ts
@@ -1,12 +1,13 @@
 import { Injectable, ConflictException, BadRequestException, NotFoundException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Users } from 'src/model/users/users';
+import { Users } from '../../model/users/users';
 import { Repository } from "typeorm"
 import * as bcrypt from 'bcrypt'
 import { CreateUserDto, LoginDto } from 'src/model/dto/user.dto';
-import { Rooms } from "src/model/rooms/rooms"
+import { Rooms } from "../../model/rooms/rooms"
 import { userdetailresponse, userResponse } from 'src/utils/types';
 import { HttpException } from '@nestjs/common';
+import { Http2ServerResponse } from 'node:http2';
 
 @Injectable()
 export class UsersService {
@@ -39,6 +40,7 @@ export class UsersService {
       isActive: data.isActive,
     }
   }
+
   async getusers() {
     const data = await this.userrepository.find({
       select: {
@@ -50,6 +52,7 @@ export class UsersService {
     })
     return data
   }
+
   async finduser(id: string): Promise<userdetailresponse> {
     let data = await this.userrepository.findOne({
       where: { id },
@@ -68,15 +71,23 @@ export class UsersService {
       isActive: data.isActive
     }
   }
+
   async remove(id: string): Promise<void> {
-    this.userrepository.delete(id)
+    const user = await this.userrepository.findOneBy({ id })
+    if (!user) {
+      throw new HttpException('user not found', HttpStatus.NOT_FOUND)
+    }
+    await this.userrepository.delete(id)
   }
+
   async login(body: LoginDto): Promise<userdetailresponse | null> {
     if (!body.password) throw new HttpException('password field is empty', HttpStatus.BAD_REQUEST)
     const check = await this.userrepository.findOneBy({ email: body.email })
     if (!check) throw new BadRequestException("email does not exist")
-    const isPasswordValid = bcrypt.compare(body.password, check.password)
+
+    const isPasswordValid = await bcrypt.compare(body.password, check.password)
     if (!isPasswordValid) throw new ConflictException("password is incorrect")
+
     return {
       id: check.id,
       username: check.username,
@@ -84,24 +95,6 @@ export class UsersService {
       groups: check.groups,
       createdgroups: check.createdgroups,
       isActive: check.isActive
-    }
-  }
-
-  async joingroups(id: string, uid: string): Promise<void> {
-    const check = await this.roomrepository.findOne({
-      where: { id }
-    })
-    console.log(check)
-    let checkuser = await this.userrepository.findOne({
-      where: { id: uid },
-      relations: {
-        groups: true
-      }
-    })
-
-    if (checkuser && check && !checkuser.groups.some(g => g.id === check.id)) {
-      checkuser.groups.push(check)
-      await this.userrepository.save(checkuser)
     }
   }
 }
